iothub_simulator/requirements.txt
azure-eventhub==5.11.0
azure-identity==1.23.1
